pyInterface: Fix several root related problems.
Fix some problems in the python bindings for the amplitudeTreeLeaf concerning
writing and reading from a root tree, adapt the branch-creating function to
utilize the possibility to set bufsize and splitlevel and add a function which
was missing to the bindings of the amplitudeMetadata. Remove a tree-related
function from the bindings of the ampIntegralMatrix, as these classes are not
stored in trees.