partialWaveFit: fix intensityErr
The 'fitResult::intensityErr' method calculated a wrong result for all
patterns that did not result in a continuous sequences of production
amplitude indices and were not starting from zero (realistically that
is all except the '.*' pattern). 'fitResult::normIntegralForProdAmp'
expects an index for the production amplitude, not an index for an
array of indices for the production amplitude.