decayAmplitude: Remove superfluous index map.
Remove an index map in the ampIntegralMatrix which was not really needed and
had probably been added to make mapping from wave name to wave index easier.
However, there was much more code caring for the correct filling and rebuilding
of the map than code actually using its contents, so it did not really make
sense to keep it.