decayAmplitude: Improve integral matrix metadata.
Revise the integral matrix metadata by correcting formatting and making several
places in the code slimmer. Also rewrite most of the error messages so that
there are no variable names in the printouts. Note that there were a few minor
interface changes which probably break python scripts based on this code, but
this will be corrected later. Also note that none of this was tested yet.