userAnalysisWorkspace: fix 'GenerateKeyfiles.sh' for 3pi and pi eta eta
The 'GenerateKeyfiles.sh' present for the 'pi- pi- pi+', 'pi- pi0 pi0',
and 'pi- eta eta' channels did not work correctly if the variable
'WAVESET_FILES' contained the full path of the wavelists. This is now
fixed by taking only the 'basename' of each wavelist when attempting to
copy and modify those files.