partialWaveFit: replace fit parameter information in likelihood
Replace the way the fit parameter information is kept inside the
'pwaLikelihood' class. Instead of storing the parameter name,
threshold and status (fixed/free) in individual vectors, use a new
class 'fitParameter' which holds the wave name and the rank, the
threshold and a 'fixed' flat, and whether it is a real or imaginary
part. The parameter name is created from these information.