tests: Fix the pylint tests for the scripts.
Fix the pylint tests for the three python scripts which have been broken by the
recent revision of the file manager and the on-the-fly binning. These fixes are
rather ugly and there is a significant amount of code duplication, but these
scripts need to be revisited and revised anyway when the automatic integral
merging for the fitting is implemented.