partialWaveFit: fix filling of 'normIntIndexMap'
The mapping of production amplitudes to entries in the normalisation
integrals was wrong for multi-rank fits for the flat wave. However,
this did not have any effect, the 'normIntIndexMap' was only used in
'normIntegralForProdAmp' which is only called from 'intensityErr'. In
that function it is assured the ranks of the production amplitudes
are equal, as the flat wave has rank -1, this could never happen.