generators: use 'nBodyPhaseSpaceKinematics' in importance sampler
Use the 'nBodyPhaseSpaceKinematics' class to calculate the
phase-space weights of events in the 'importanceSampler' class
instead of reimplementing the calculation. The
'nBodyPhaseSpaceKinematics' class also provides the functionality to
calculate the Lorentz-vectors of the final-state particles. Change
the ordering of variables in BAT to have the same convention for the
variables.