pyInterface: fix a seg fault in 'genPseudoDataImportanceSampling'
Fix a segmentation violation the happened when
'genPseudoDataImportanceSampling' was finishing. Internal BAT
histograms were created with the event file taking ownership. When
closing the event file they were no longer available, and the
'importanceSampler' class could not be deleted in a safe way. Now
change to ROOT's root directory after opening the event file, so that
everything should be okay.