storageFormats: binning map during merge
While merging event data the binning map was not set if all binning maps
were requested to be equal, as in the very end it was overwritten with
an empty map. As before the binning map for this case is set in the very
beginning. Only in case of a new/merged binning map, it is reset in the
end.