resonanceFit: fix branchings for new naming scheme
For the fit using branchings the wave names need to be split into the
quantum number and the decay part. Until now this was done based on the
old naming scheme (first seven characters are the quantum numbers, the
rest the decay chain), and is now replaced by the new naming scheme
(quantum numbers and decays are separated by the equal sign).

For the fits using only the couplings this has no effect.