resonanceFit: support different mass binnings
Add support for different mass binnings in the different bins used in
the resonance fit. The mass binnings can be different in number,
position, and widths. For this it was neccessary to add one dimension
to basically every variable storing information on the mass binning.
Caching of intermediate results needs to ensure to set the value for
the appropriate bin(s) only, as the same 'idxMass' variable might
indicate different masses.