resonanceFit: different wave names in bins
Allow the usage of different wave names in different bins. The name
used for the title is still the name given by the 'name' option, but
additional names can be given with the 'alternativeNames' option. It
is ensured that no name can appear multiple times. While retrieving
information from the fit result all names are tried before giving up.