resonanceFit: simplify alternative wave name handling
Simplify the handling of the alternative wave names. Assuming that
all mass bins within one one contain the same waves, the wave names
to use while extracting information from the fit result can be
created once in the beginning of 'readFitResultMatrices', and only
those names can be used later on, removing the duplicate search for
a match in the alternative wave names.