resonanceFit: reduce number of arguments and member variables
Reduce the number of arguments for the 'init' methods of the 'channel'
and 'component' classes: the 'nrBins' and the 'sameMassBining'
variables can be deduced from the other arguments. In the 'model'
class do not store the 'waveNames' array as a member variables, this
is only used during the initialization, so it can be used from the
arguments.