resonanceFit: use certain waves in individual bins
Depending on the wave name, individual channels can now contribute
only in a subset of bins, and do not have to contribute in all. The
'channel' class has a new member that stores the list of bins, in
the 'model' class, it is ensured that also the anchor wave can vary
from bin to bin. This is a requirement to fit to two channels at the
same time if different parameterisations are to be used in the two
channels, e.g., backgrounds that should be different for the channels,
but should be the same in all bins of one channel.