Test: improve reporting of script for test MC
Fix misbehavior of sed line on BSD-like systems, as reported by
Sebastian: Turns out that with GNU sed
> sed -i '' <command> <file>
gives error
> sed: can't read <command>: No such file or directory
but works fine with BSD sed, whereas
> sed -i'' <command> <file>
works fine with GNU sed, but gives error
> sed: 1: "<file>": invalid command code <first letter of file>
with BSD sed. So, this is a no-win situation. I therefore reverted to
creating the backup files and removing them immediately.