partialWaveFit: allow partial calculation of evidence components
Allow the partial calculation of evidence components by first
initializing the result vector for the four components with negative
infinity and then overwriting the individual components if they can be
calculated.

The first component should always be calculable, the other three require
the covariance matrix. The third component in addition requires the
integral matrices.