pyInterface: search for amplitude metadata in 'printMetadata'
Do not require that 'printMetadata' is called with an argument
specifying which amplitude to print, but loop over the keys of the ROOT
file, and print all amplitude metadata objects that are found. This
covers the typical case of one amplitude per file, but should also be
working if multiple amplitudes are stored in the same file.