generators: reduce code duplication in 'initDecayAmplitudes'
Reduce the amount of code that is duplicated in the two
'initDecayAmplitudes' functions. Add a third version of this method that
is templated and depending on whether it is called with a constant
vector for the names of the production particles or not uses this vector
or obtains the required information from the first valid decay
amplitude.