generators: move the code to calculate the X Lorentz vector
Move the code calculating the Lorentz vector of X from the Lorentz
vector of the beam, the masses of the involved particles and the
Mandelstam variables to a static class function inside the
'diffractivePhaseSpace'. The same code was also used in the
'importanceSampler' class, so there the new function is now called.