decayAmplitude: use isobar mass-dependencies in wave set generator
Use the isobar mass-dependencies that are read from the decay config
file in the wave set generator while create the wave set. For each
isobar decay check the matchings decay modes for the mass-
dependencies to be used. If no mass-dependence is specified the
default is used by not setting anything (currently:
'relativisticBreitWigner'). Also add a check that the names of the
generated waves are unique, otherwise keyfiles of waves might be
overwritten in one 'generateWaveSet' run.