userAnalysisWorkspace: mass dependencies in wave-set generator for 3pi
Specify the isobar mass dependencies in the three pion channels for
the wave-set generator. Currently the 'sigma0' and the 'f0(980)0' use
non-standard mass dependencies. For the moment only a single
parameterisation for the 'f0(980)0' is used, this should be extended
again once the wave names when using different mass dependencies are
also different.