pyInterface: fix name of COMGEANT output file in genpw
The name of the COMGEANT output file was created from the old '.evt'
name. As we are not using the '.evt' files for a long time, the
replacement failed and the COMGEANT output file was immediatly
overwritten by the ROOT output file.

Make sure that the output file name specified on the command line has
the '.root' file extansion as otherwise the replacement will not work.