generators: replace 'generatePhaseSpace' by Python script
Remove the C++ executable 'generatePhaseSpace', and rename the current
'genpw' script to 'generatePhaseSpace.py' (while running 'make' a link
without the '.py' is created). Adjust the test script accordingly.

In contrast to the C++ executable the Python executable does not
understand the option '-r' to specify the reaction file, but instead
uses a positional argument.

It seems to be necessary to remove the content of the 'build/bin/'
directory for this renaming to work properly.