test: fix 'testMC.sh'
Fix the arguments passed to 'genPseudoData' in 'testMC.sh'. Following
e9aaa498d104c5f5ba991cd2daa05e9aff1119a2 the phase-space integral is by
default taken from the fit result and no longer needs to be specified.
So the integral is no longer a positional argument, but instead has to
be given with the '-i' option, and in the case here this is also
necessary, as the used fit result does not contain any matrices.