resonanceFit: restructure caching and calculation of components
Restructured the way the values for the components are calculated and
cached. The 'val' function with the previous signature has been made
non-virtual in the 'component' class. It takes care of getting and
putting the value from the cache, the actual computation is performed
in a new 'val' function not having the cache in the argument list.
This removes a lot of code duplication required for the caching in
each of the various components.