resonanceFit: more flexible caching of components
The caching of intermediate results for components was not efficient
when using different mass binnings in the individual bins of the fit.
In such a case the values were only stored for the current bin. This
has now been extended. A new member '_binsEqualValues' has been
introduced which indicates for each bin all other bins in which the
component should yield the same value given the same input. By
default for each bin this is only the current bin (and this is used
for the t' dependent background parameterizations). The
'initBinsEqualValuesFromMassBinning' function checks the mass binning
for equal number of bins and equal center of the mass bins to decide
which bins can be considered to have equal values.