resonanceFit: store information for fit quality in a map
No longer store the information on the fit result in individual values
for 'chi2', 'ndf' and so on. Instead the 'minimize' method now returns a
map from strings to doubles which contain these information. During
writing all entries of this map are dumped to the YAML file, during
reading all entries in the YAML file are stored in this map. A couple of
additional information is stored in case of the ROOT fitter like the
(explicit) status of the covariance matrix, the estimated distance to
minimum and so on.