resonanceFit: signal if maximal number of calls was exceeded
Add a key 'continue' to the map 'fitQuality' if the maximum number of
allowed function calls was exceeded. This can be used by external
programs to determine more reliably whether the fit has finished, or was
stopped due to this. Previously it might have happened, that one stage
finished within the allowed number of function calls, but the next could
not be started (as the allowed number of function calls for the new step
would be 0). In such a case the minimizer status would not be 4, but
would be the proper status of the minimization.