resonanceFit: prevent crash if error is 'nan'
Sometimes the minimizer returns 'nan' as the value of an error. Some
systems, e.g. SLC6, cannot convert the 'nan' string read from the YAML
file in such a case back to a floating-point number. Now perform this
conversion manually. This is in any case not really important, the error
is not used for anything in the current version.