resonanceFit: rename source file of executable
Rename the source file of the executable to 'resonanceFitExe.cc'
(instead of 'resonanceFit.cc').

The plan is to add a couple of functions taking care of basics around
the resonance fit to a new source file 'resonanceFit.cc', and call
those functions from a Python script instead of having the C++
executable. While developing this, the old executable needs to be
kept around, but needs to give up its name.