resonanceFit: move production amplitudes to 'data' class
Move the production amplitudes and corresponding covariance matrices
required for the fit to the 'data' class. The corresponding members
of the 'massDepFit' and 'function' classes have been replaced by
using the new class instead.

The preparation of the production amplitudes and covariance matrices
has been moved from the 'function' class to the anonymous namespace
in 'massDepFit.cc'.