resonanceFit: start concentrating data in a single class
Add a new 'data' class. For the moment this class only holds the
number of centers of the mass bins for each bin. The corresponding
members of the 'massDepFit' and 'function' classes have been
replaced by using the new class as far as possible. This should help
to only have the data at one place.