resonanceFit: restructure variables for plotting
Restructure the arrays used to store the variables used for plotting.
Use separate arrays from those used and processed for the fit. Store
a pair of doubles for the value and its error. Do this for the
plotting of the fit results used in the fit, as well as those used to
indicate systematic effects.