resonanceFit: systematics plotting with several bins
Enable the drawing of systematic uncertainties when fitting in
several bins. For this, the 'systematics' part in the yaml file is
now required for each element of 'fitresults' if systematics are to
be shown. The arrays of pair introudced in the previous commit now
contain the lowest and highest value found for one entry (and no
longer value and error).