resonanceFit: start concentrating information in a single class
Add a new 'information' class. The purpose of the class is to store
information required to load everything that is required by the fit,
like the file names of the fit results, the waves to be used, and so
on. These information are currently stored in the 'massDepFit' class.

For the moment the class is only used to store the information of the
bins: the file name for the fit result used in the fit, the mean
value of tPrime, the rescale factor for the errors, and the file
names of the fit results used for the systematic errors. The
corresponding members of 'massDepFit' have been removed. Reading and
writing the information from and to the configuration file is now
performed in the anonymous namespace in 'massDepFit.cc' instead of
private functions in the 'massDepFit' class.