resonanceFit: remove '_sameMassBinning' member from 'massDepFit'
It is only used for the initialization of  the final-state mass-
dependence and can be obtained from the mass bin centers in there. In
'massDepFit' the consistency of the different mass binnings is no
longer relevant, and also no longer checked. In 'fsmd' a similar
mechanism as introduced for the components in
5d132c28ce43dbcb653509b23aab660c2f95903b has been added.