resonanceFit: move 'freeparameters' in config file
Move the 'freeparameters' section in the configuration file from the
'input' section to the root. There might be minimizers that have no
use for different ordering of freed parameters. Also remove the
'_freeParameters' member from the 'massDepFit' class. Move the
functions reading and writing the release order from and to the
configuration file to the anonymous namespace in 'massDepFit.cc'.