resonanceFit: change initialization of 'components'
Change the wave the components are initialized. All code reading the
YAML configuration is removed from the 'component' and inheriting
classes. The parsing of the configuration file is instead performed
in 'massDepFit.cc'. The same is done for writing the YAML file.

The 'component' classes no longer have a 'init' function, but are
fully initialized in the constructors. The 'channel' class has been
made a subclass of 'component'. The decay channels now need to be
created before calling the component constructor.