resonanceFit: change printing of components
Print all parameters, names and values in the 'print' function of the
'component' class, in the class-specific functions only the class-
specific information are printed.

The 'print' method, and thus also the 'operator<<' overload always
printed a new line after printing the information contained in the
classes. This was an issue if this information is printed from, e.g.,
Python, where an extra empty line would show up.

The 'print' function now has an extra argument determining whether to
print the new line or not, the 'operator' overloads do not print it.