resonanceFit: change initialization of 'model'
Objects of 'model' class are now fully initialized in the
constructor, a separate call to 'init' is no longer required. For
this, the '_anchor...' members of 'massDepFit' have been removed, and
are now local in the function reading the model. The functions reading
the information for the model have been moved to the anoynmous
namespace in 'massDepFit.cc' from the 'massDepFit' class.