decayAmplitude: get parent for wave name from 'massDependence'
Get the string used to describe the parent of a decay in wave names
from the 'massDependence' using the information from the decay
vertex. Use different tags for the various dependencies. The
'relativisticBreitWigner' mass dependence has no tag, as was the
case before. The 'flat' and 'binned' mass dependencies now contain
quantum numbers, while all others contain the name of the parent
state.