test: update 'testMC.sh' for updated wave-set generator
Use the particle decay tables in the 'testMC.sh' script instead of
relying on the 'GenerateWaveSet.sh' script to take care of creating
the desired versions of the keyfiles. The update of the wave-set
generator changed some wave names. This required to update the
reference fit result and to use not the original wave list, but one
with a modified name for the f0(980) waves.