pyInterface: fix 'getBestFitResultFromFile'
Fix the 'getBestFitResultFromFile' for cases when no 'massBinCenter' is
given as an argument. In this case 'getBestFitResultFromFile' should get
the best fit result in the file.

This was never the case since the introduction of this function in
d15d0108828492f45ff887a07d3af26ad2d8ae76. The comparison of the
likelihoods was only performed if the current fit result had a
different mass bin center from the best fit result encountered so far.