pyInterface: set ROOTPWA git has for 'ampIntegralMatrixMetadata'
While calculating the integral, set the ROOTPWA git hash to the current
version. This means that the revision of the current executable is
stored in the metadata, the revision of the individual amplitudes can be
obtained by checking the amplitude files and comparing the content of
these files with the amplitude hash stored in the metadata.