pyInterface: add option to fit without normalization of amplitudes
Re-add the option to perform the partial-wave fit without the
normalization of the amplitudes to the Python scripts. This
functionality has not been taken over from the previous C++
executables. A new command-line flag '--do-not-normalize-amplitudes'
has been added to both fitter scripts. The resulting option is used
while setting up the likelihood replacing the previously hard-coded
always normalize. Set the number of events to normalize to
accordingly in the fit result.