pyInterface: add converter 'boost::multi_array' to 'numpy.array'
Add some converters from 'boost::multi_array' to 'numpy.array' (which
seems to be the library of choice for multidimensional arrays in
Python). All converters required for the resonance fit are added,
that are for elements of type double, complex, TMatrixT, pairs of
size_t and pairs of double, and strings.