resonanceFit: allow to release individual couplings and branchings
Allow to release individual couplings and branchings by specifying
the full name of the paramater (without '__real' or '__imag'). For
the couplings allow the release all couplings of a single bin.

This means that the string of free parameters can no longer be split
at ',' characters as those are part of the wave names according to
the new naming scheme.