decayAmplitude: reduce message by 'constructDecayTopology'
Reduce the number of messages shown whenever the file manager is loaded
by no longer printing information messages in
'waveDescription::constructDecayTopology'. The messages printed to
indicate starting and finishing the construction are only printed if the
debug flag is set. Corresponding messages have instead been added
directly to the file manager and the likelihood, which are the main
users of 'constructDecayTopology', to show how many waves are known in
the file manager or are used in the fit.