buildSystem: allow older Boost versions
In commit d0279c250cea557146f9c5427a2de95d665f01ba the required Boost
version was increased from 1.50.0 to 1.56.0 because the build system of
Boost was changed, and out script to build the library needed to be
adapted. However, nothing in Boost itself prevents to us from using an
older version. Undo this change so we can still use the version 1.54.0
available on Travis (Ubuntu Trusty Thar) and do not have to install from
source.