decayAmplitude: intermediate mass dependence for static functions
Add an intermediate class for the mass dependencies. This intermediate
class takes as a template parameter the class that inherits from it, and
thus can provide static functions for a number of classes without
duplicating too much code. In this case the intermediate class provides
static methods to  obtain the name of the mass dependence and to create
a new object. It also overrides the virtual function 'name' of the old
'massDependence' base class, such that this 'name' returns a value
equal to the static name.