decayAmplitude: rename 'parentLabelForWaveName'
Rename the function 'parentForWaveName' of the mass dependence classes
to 'parentLabelForWaveName'. Implement this function in the base class,
because a lot of classes actually do not need any specialization. In the
'flat' and 'binned' case do not use the hard-coded names, but instead
those from the 'name()' function.

This commit also ensures that the name of the mass dependence is the
same across keyfile and wave name.