decayAmplitude: generalize 'createMassDependence'
Remove the hard-coded names for the mass dependencies in the
'createMassDependence'. Instead pass a list of classes to a templated
function. To name of each of these classes is compared to the requested
name, and an object of the according type is created. A nullptr is still
returned if the name could not be resolved.