resonanceFit: allow completely different structures of bins
Add the possibility to use bins that have completely different
structures. This ranges from just different mass ranges used for a
wave to a completely different set of waves used for each bin. This
required quite some restructuring of quite a lot:

The 'rpwa::resonanceFit::input::wave' class has been made a subclass
of '...::input::bin'; alternative wave names are no longer allowed,
it is assumed that all mass bins of a bin use the same wave names, so
it is possible to specify the exact wave name. The number of waves
might now differ from one bin to another.

The constructor of the class '...::component::channel' now requires a
list wave indices in each bin, it recognizes a special value if a
wave does not exist in a bin. The list of bins for a channel is
automatically created from this, and no longer needs to be specified.

It is now also required to specify the anchor wave and anchor
component names for each bin.