partialWaveFit: Implement getFitResultsFromFiles and fill functions without matrices
Implement loading of fit results from files and sorting them in
multibins.
Implement python interface for this function.

Implement specialized classes of the fill function, which allow to fill
a fitResult without storing the covariance and integral matrices.
Additionally, implement the corresponding copy constructors and use the
fill function in the copy constructors such that filling is now
collected in one place in the code.