partialWaveFit: Implement getFitResultsFromFiles and fill functions without matrices
Implement loading of fit results from files and sorting them in
multibins.
Implement python interface for this function.

Implement specialized classes of the fill function, which allow to fill
a fitResult without storing the covariance and integral matrices.
Additionally, implement the corresponding copy constructors and use the
fill function in the copy constructors such that filling is now
collected in one place in the code.

partialWaveFit: Fixed bug and flaws in `fill` function

The `_waveNames` was not build when neither the `normIntegral` nor the
`acceptedNormIntegral` was given.
	This bug is fixed.
The `fill` function form another `fitResult` prints lots of warnings
when filled from another `fitResult` without one or more matrices.
	This warnings do not appear anymore
The `fitParCovMatrixIndices` is only required when `fill` is called with
a `fitParCovMatrix`. This is inconsistent with the behaviour of the
`saveSpace` flag, where the `fitParCovMatrixIndices` are always stored.
	Require `fitParCovMatrixIndices` for all `fitResults`.
The specialized `fill` function, which are used to fill a `fitResult`
from another `fitResult` handled an empty prodAmps vector to the `fill`
function and copied the `_prodAmps` variable directly to avoid
additional conversion.
This caused a warning message every time fill was called.
	Now, there is a dedicated signature of the `fill` function for this.