decayAmplitude: change name and keyfile name for 'f0980FlatteBesII'
Change the name of the 'f0980Flatte' mass dependence to
'f0980FlatteBesII", and also require that name in the keyfiles. The name
used in the keyfile is now equal to the name of the class, the old name
is still accepted, but prints a warning. The 'BesII' has been added to
allow for future distinction to other Flatte parameterisations.