pyInterface: fix error message in file manager
An error message in the file manager still expected the 'keyFiles'
dictionary to contain the keyfile name and (in case of keyfiles that are
expanded) the index in the expansion. 'keyFiles' does not contain this
index for some time already, so also remove this from the error message.

Without this patch, the error message only printed single characters,
which is quite useless.