pyInterface: Additional functionality of the multiBin class
Implement hash function for the multiBin class, such that
it can be used as key in a dict.
Additionally, the printing of the multiBin class is improved.
Implement fetching of a sub-bin.
Implement 'A in B' operator if bin A is contained in bin B.
Remove restriction that a multiBin must have at least one variable
(this restriction, if needed, should be implemented in the config file)
Remove constrain that a the boundaries of a multi-bin cannot be equal,
because this is needed if we have only the bin-center information
available (e.g. old version of fitResult class).