Fix ROOT I/O after renaming to 'multibinBoundaries'
Fix commit 0c6259e82f6ad38d63135fb0fa8e109472a8c64b. The version in
'ClassDef' statements is now increased, and read rules to set the
content of 'multibinBoundaries' to the content of the previous
'binningMap' are added.

In principle the same would have to be applied for 'fitResult', but
since adding the binning map to the fit result and renaming it have been
merged at basically the same time, this is ignored.