buildSystem: Fix python problem in travis
There seem to be a change in the travis system, such that another python
version is now used. As a result, the libpython.so does not exist
in the LIBDIR. Thus, the symlink to libpython.so was broken. Use 'pyenv'
to force the system python and remove all remainders of the workaround.