decayAmplitude: extend 'createMassDependence'
Extend the 'createMassDependence' function to also work with mass
dependencies that require additional settings (like the binned mass
dep.). In these cases a second static 'Create' function with a 'Setting'
argument should be added to the mass dependence class, which takes care
of extracting the required information and then calls the previous
'Create' function. In 'createMassDependence', depending on whether the
mass dependence can be constructed without any argument or not, either
the 'Create' function is called without arguments, or with the provided
'Setting' object. Calling 'createMassDependence' for a mass dependence
that would require additional settings without a 'Setting' object will
throw at runtime.