storageFormats: change hash of 'amplitudeMetadata'
Use the content hashes of the used event metadata and the wave
description when calculating the hashes for amplitude files.

The wave descriptions (keyfile contents) are used in order to be able to
distinguish between several waves that have equal amplitudes for all
events. This might happen, e.g., for waves with the binned
mass-dependence if the invariant mass of X for all events is below the
lower bound of the bins.

This is not sufficient if two event files for which all events have an
equal amplitude for each event in both files for one wave, and the two
event files have the same number of events, so also use the hash of the
event information.

Revert 9fce75b9d027c8f575aeec77a0092abb2f6e64d2 now that the hashes are
unique.