partialWaveFit: fix production amplitudes for 'empty' waves
If the decay amplitude for a wave is zero for all events, and if the
corresponding elements in the normalization matrix with acceptance are
also zero, then this wave does not contribute to any quantity calculated
during the fit. This might spoil the errors that Minuit calculates, so
fix the production amplitudes corresponding to these waves to zero.