buildSystem: prepare for Boost >= 1.67.0
Since Boost 1.67.0 the Boost.Python libraries include the Python version
in their names (i.e. libboost_python27.so instead of just
libboost_python.so). Cmake has added support for this since version
3.11.0. This requires to search for the Boost component, e.g., python27
(instead of just python). Cmake will fall back to the name without the
version if the library with the version in its name is not found.

The order of finding Boost and Python had to be changed, such that first
Python is looked for, and its version number can be used when Boost is
searched.

Also make use of the OPTIONAL_COMPONENTS option that FindBoost.cmake has
instead of having our own hack to require optional components.