pyInterface: Generate amplitude filepaths dynamically in `fileManager`
Before, the full path to all amplitude files was generated once and
stored in the `fileManager`.
For large modes and fine data binning, the `fileManager` object in
memory becomes very large > 1GB and takes some time to load.
Now, the amplitude filepaths are generated dynamically on request.
This saves memory and is approx. 2x faster than loading the filepaths
for 1000 waves and 500 data bins.