partialWaveFit: Fix bug for reading `complexMatrix` from ROOT file
When reading a `complexMatrix` with size 0 form a ROOT file, the
internal matrix is not changed.
If the `complexMatrix` object had 0 size before, e.g. a new
`complexMatrix`, this is fine.
However, if the `complexMatrix` object was not 0, e.g. if the previous
matrix in the tree was not 0, the `complexMatrix` object will keep the
data from the previous one.