buildSystem: bump up versions of some dependencies
Increase minimum versions of CMake to 3.11 and of Boost to 1.67
(changed handling of Boost python module). Increase minimum version of
ROOT to 6.08 and of Python to 2.7; do not allow Python 3. Enable CMake
module for NLopt to also work with CMake build of NLopt introduced
with version 2.5.0. Update and improve README. Adapt and update Travis
environment. Improve coding style.